refactor: extract backup manager and timer
